requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
twilio>=8.0.0
